pkg/trace/quantile: move to pkg/trace/stats/quantile
This commit is the result of running:

	gomvpkg -from github.com/DataDog/datadog-agent/pkg/trace/quantile -to github.com/DataDog/datadog-agent/pkg/trace/stats/quantile
